Hide VPN internals
Shield concrete controller behind interface.
