+++ cross/android/io/partout/PartoutVpnServiceRuntime.kt
@@ -17,6 +17,8 @@ import android.util.Log
 import io.partout.models.TaggedProfile
 import io.partout.models.TunnelSnapshot
 import io.partout.models.TunnelStatus
+import io.partout.vpn.JNITunnelController
+import io.partout.vpn.TunnelControllerDelegate
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -139,7 +141,7 @@ class PartoutVpnServiceRuntime(
         engine.onSnapshot(snapshot)
     }
 
-    override fun shouldDisconnect(controller: JNITunnelController) = launchCommand {
+    override fun shouldDisconnect(controller: NativeTunnelControllerJNI) = launchCommand {
         if (controller != this.controller) { return@launchCommand }
         stopTunnel()
         stopService()
@@ -367,7 +369,7 @@ class PartoutVpnServiceRuntime(
 
     //region Engine
     interface Engine {
-        suspend fun start(intent: Intent?, controller: JNITunnelController, profileJSON: String)
+        suspend fun start(intent: Intent?, controller: NativeTunnelControllerJNI, profileJSON: String)
         suspend fun stop()
         suspend fun readLastProfile(): String
         suspend fun writeLastProfile(json: String)
