--- cross/android/io/partout/vpn/VpnServiceApplying.kt
@@ -1,11 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Davide De Rosa
-//
-// SPDX-License-Identifier: GPL-3.0
-
-package io.partout.vpn
-
-import android.net.VpnService
-
-interface VpnServiceApplying {
-    fun apply(logTag: String, builder: VpnService.Builder): Boolean
-}
